refactor(cli): extract operator input parsing from ideas command
ideas.py (2770 lines) interleaved its JSON payload / fixture parsing concern
(trade-idea / market-snapshot / replay-candle loaders, the InputPayloadError
hierarchy, and the _error_context helper) with command registration and
handlers. Move the whole parsing concern into cli/commands/ideas_input.py
(2770->2484) and re-export the handler-facing names.

The parsers are pure (JSON -> domain object, raising InputPayloadError
subclasses) and reference nothing else in ideas.py. One stdin test that patched
the now-moved module-global sys now patches sys.stdin directly. 253 CLI tests pass.
